+++ plugins/O-MicrotonalSampler/Source/MicrotonalSamplerSound.h
@@ -15,8 +15,6 @@
 class MicrotonalSamplerSound : public juce::SynthesiserSound
 {
 public:
-    MicrotonalSamplerSound() {}
-
     bool appliesToNote   (int) override { return true; }
     bool appliesToChannel(int) override { return true; }
 };